+++ backup_data.py
@@ -1,0 +1,31 @@
+# Old
+#monthly_rate = 12 * (interest_rate + 1)**(1/12)
+#monthly_rate = (1 + (interest_rate/12))**12 - 1
+#print("Monthly Rate: ",monthly_rate)
+
+#mortgage_rate = 2661.21
+
+
+#monthly_rate = interest_rate / 12
+
+"""
+i,p = interest_payment(loan_amount,interest_rate,loan_term,mortgage_rate)
+
+print(i)
+print(p)
+
+i,p = interest_payment(loan_amount-p,interest_rate,loan_term-(1/12),mortgage_rate)
+
+print(i)
+print(p)
+"""
+
+
+"""
+plt.figure()
+plt.plot(months, mpp)
+plt.xlabel("Months")
+plt.ylabel("Principal Payment")
+plt.grid(True)
+plt.show()
+"""
